refactor(FirebaseAuthContext): convert googleUser to appUser
The GoogleAuth User object that is returned after auth is very large
and and containes a lot of extraneous information that the app
does not need. Using this object also created difficulty with type
checking in tests. As a result I refactored the FireBaseAuthContext
component to extract a more simplified version of the google user data
and created a User type that can be used throughout the application.

After the refactor I went through the components that relied on the
original google user type and updated it to the simplified user type.