feat(FirebaseAuthContext): add user auth database functionality
added logic to handle adding user to the firebase firestore if on
google auth they do not already exist in the firestore. If they do exist
in the firestore then the firestore user document is returned instead
of the google auth document. This ensures that the user data provided
by the FirebaseAuthContext reflects the data in the firestore for that
user.