refactor(UserAuth): to use React Query over FirebaseAuthContext
The google auth listener that is setup in the FirebaseAuthContext
did not provide enough functionality for me to reliably get my
user data into the components that I needed. This was realized when
a user fills out the new registration survey. After this occurs I
need to update the firestore database and then refetch the user from
the firestore to make sure that the user state stays in sync with the
application. React Query works very well with this because we can
invalidate chosen queries which initiates an auto refetch. As a result
I refactored the application to use React Query instead of the prior
Auth Context provider.